backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
openai==1.3.7
werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2